solutions/day-1/destination-advisor/requirements.txt
agent-framework
python-dotenv
uvicorn
